modify app server module file, import methods refactor
